pyproject.toml
[tool.poetry]
name = "gloard"
version = "0.1.0"
description = "Tame the Gitlab productivity frendziness of your team in an easy-to-use dashboard"
authors = ["Valentin Famelart <valentin.famelart@invenis.co>"]
license = "MIT"

packages = [
    { include = "src" },
]

[tool.poetry.scripts]
server = "src.server:run_server"

[tool.poetry.dependencies]
python = "^3.9"
tartiflette-asgi = "^0.9.0"
uvicorn = "^0.12.2"
environs = "^9.3.0"
python-gitlab = "^2.6.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
flake8 = "^3.8.3"
isort = "^5.6.4"
requests = "^2.25.0"
ipdb = "^0.13.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
